A Concordance Project
Using Collection classes build a concordance.
A concordance lists every word used in a document and where each
instance is.
For example a biblical concordance would show you every use (page and
line) of the word “beget”.
Use a hash for the words and a linked list for every instance.